+++ src/builders/aggregators/abstractAggregator.ts
@@ -1,19 +1,26 @@
+/* eslint-disable max-len */
 import { AbstractColumn, Column } from '../../columns/column';
 import ColumnType from '../../columns/types/columnType';
 import { AbstractTable } from '../../tables';
 import { ecranate } from '../../utils/ecranate';
 
+// eslint-disable-next-line max-len
 export default class Aggregator {
   protected _fields: Array<string> = [];
   protected _table: AbstractTable<any>;
 
-  public constructor(table: AbstractTable<any>) {
+  public constructor(table: AbstractTable<any>, partial?: {[name: string]: AbstractColumn<ColumnType<any>, boolean, boolean, AbstractTable<any>>}) {
     this._table = table;
-    this._fields = this.generateSelectArray(this._table.tableName(),
-      Object.values(this._table.mapServiceToDb()));
+    if (!partial) {
+      this._fields = this.generateSelectArray(this._table.tableName(),
+        Object.values(this._table.mapServiceToDb()));
+    } else {
+      this._fields = this.generateSelectArray(this._table.tableName(),
+        Object.values(partial));
+    }
   }
 
-  protected generateSelectArray = (table: string, columns: AbstractColumn<ColumnType>[]) => {
+  protected generateSelectArray = (table: string, columns: AbstractColumn<ColumnType>[], id?: number) => {
     const selectFields: string[] = [];
 
     columns.forEach((field: any) => {
@@ -23,7 +30,7 @@ export default class Aggregator {
         selectFields.push('.');
         selectFields.push(ecranate(field.getColumnName()));
         selectFields.push(' AS ');
-        selectFields.push(ecranate(`${table.replace('.', '_')}_${field.getColumnName()}`));
+        selectFields.push(ecranate(`${field.getAlias()}${id ? `_${id}` : ''}`));
         selectFields.push(',');
       }
     });