+++ src/builders/highLvlBuilders/abstractRequestBuilder.ts
@@ -5,7 +5,8 @@ import BaseLogger from '../../logger/abstractLogger';
 import { AbstractTable } from '../../tables';
 import { ExtractModel } from '../../tables/inferTypes';
 
-export default abstract class TableRequestBuilder<TTable extends AbstractTable<TTable>> {
+export default abstract class TableRequestBuilder<TTable extends AbstractTable<TTable>,
+TPartial extends {[name: string]: AbstractColumn<ColumnType<any>, boolean, boolean, TTable>} = {}> {
   protected _table: TTable;
   protected _session: ISession;
   protected _mappedServiceToDb: { [name in keyof ExtractModel<TTable>]
@@ -25,16 +26,26 @@ export default abstract class TableRequestBuilder<TTable extends AbstractTable<T
     this._logger = logger;
   }
 
-  public all = async (): Promise<Array<ExtractModel<TTable> | undefined>> => {
+  public all = async () => {
     const res = await this._execute();
     return res;
   };
 
-  public first = async (): Promise<ExtractModel<TTable> | undefined> => {
+  /**
+   * Current function will return an element only if response is of length 1
+   * If there are more or less than 1 element, will throw an Error
+   */
+  public findOne = async () => {
     const executionRes = await this._execute();
-    // TODO add checks for undefined or null
-    return executionRes[0];
+    if (executionRes.length > 1) {
+      throw new Error('Request contains more than 1 element');
+    } else if (executionRes.length < 1) {
+      throw new Error('Request contains less than 1 element ');
+    } else {
+      return executionRes[0];
+    }
   };
 
-  protected abstract _execute(): Promise<Array<ExtractModel<TTable> | undefined>>;
+  // eslint-disable-next-line max-len
+  protected abstract _execute(): Promise<Array<[keyof TPartial] extends [never] ? ExtractModel<TTable>: ExtractModel<TPartial>>>;
 }